# despecialize VertexModel/EdgeModel