--- a/src/NetworkDynamics.jl
+++ b/src/NetworkDynamics.jl
@@ -125,11 +125,9 @@ function reloadfaces!()
 end
 # NetworkDynamics.reloadfaces!()
 
-using PrecompileTools: @setup_workload, @compile_workload
-@setup_workload begin
-    @compile_workload begin
-        # include("precompile_workload.jl")
-    end
+using PrecompileTools: @compile_workload
+@compile_workload begin
+    include("precompile_workload.jl")
 end
 
 end