--- a/src/component_functions.jl
+++ b/src/component_functions.jl
@@ -582,8 +582,8 @@ dispatchT(T::Type{<:EdgeModel}) = EdgeModel
 # TODO: introduce batchequal hash for faster batching of component models
 batchequal(a, b) = false
 function batchequal(a::ComponentModel, b::ComponentModel)
-    compf(a)  == compf(b)  || return false
-    compg(a)  == compg(b)  || return false
+    compf(a) === compf(b)  || return false
+    compg(a) === compg(b)  || return false
     fftype(a) == fftype(b) || return false
     dim(a)    == dim(b)    || return false
     outdim(a) == outdim(b) || return false