The modes are super specialized with concrete types. However they are only ever passed arround at the toplevel code. Here i want to check if lower compile times are noticable and worth the increased dynamic dispatches.