@hexaeder 2025-03-20
Timings:
I had a separate precompilation for the mtk model build and instantiation, so those timings did not change with the precompilation workload in nd mtk ext.

old (with specialization)
```
# ────────────────────────────────────────────────────────────────────────────────────
#                                            Time                    Allocations
#                                   ───────────────────────   ────────────────────────
#         Tot / % measured:               133s /  99.2%           12.4GiB /  99.3%

# Section                   ncalls     time    %tot     avg     alloc    %tot      avg
# ────────────────────────────────────────────────────────────────────────────────────
# Bus building (simplify)        1    46.0s   35.0%   46.0s   3.35GiB   27.2%  3.35GiB
# Initialize                     1    26.8s   20.4%   26.8s   2.46GiB   19.9%  2.46GiB
# Solve system                   1    19.0s   14.4%   19.0s   2.27GiB   18.4%  2.27GiB
# Import                         1    13.4s   10.2%   13.4s    790MiB    6.3%   790MiB
# Solve powerflow                1    11.3s    8.6%   11.3s   1.65GiB   13.4%  1.65GiB
# Line instantation              1    8.21s    6.2%   8.21s   1.18GiB    9.5%  1.18GiB
# Network building               1    5.67s    4.3%   5.67s    622MiB    4.9%   622MiB
# MTK Model creation             1    866ms    0.7%   866ms   33.9MiB    0.3%  33.9MiB
#   MTK model instntation        1    855ms    0.7%   855ms   33.8MiB    0.3%  33.8MiB
#   MTK model definition         1   10.1ms    0.0%  10.1ms    106KiB    0.0%   106KiB
# define callback                1    214ms    0.2%   214ms   8.98MiB    0.1%  8.98MiB
# ────────────────────────────────────────────────────────────────────────────────────
```
New (without specialization)
```
────────────────────────────────────────────────────────────────────────────────────
#                                            Time                    Allocations
#                                   ───────────────────────   ────────────────────────
#         Tot / % measured:               126s /  99.2%           12.2GiB /  99.3%

# Section                   ncalls     time    %tot     avg     alloc    %tot      avg
# ────────────────────────────────────────────────────────────────────────────────────
# Bus building (simplify)        1    43.1s   34.5%   43.1s   3.47GiB   28.7%  3.47GiB
# Initialize                     1    21.8s   17.5%   21.8s   2.22GiB   18.4%  2.22GiB
# Solve system                   1    21.0s   16.8%   21.0s   2.27GiB   18.8%  2.27GiB
# Import                         1    13.7s   11.0%   13.7s    790MiB    6.4%   790MiB
# Solve powerflow                1    10.7s    8.6%   10.7s   1.61GiB   13.4%  1.61GiB
# Line instantation              1    8.33s    6.7%   8.33s   1.20GiB   10.0%  1.20GiB
# Network building               1    4.90s    3.9%   4.90s    494MiB    4.0%   494MiB
# MTK Model creation             1    990ms    0.8%   990ms   33.9MiB    0.3%  33.9MiB
#   MTK model instntation        1    981ms    0.8%   981ms   33.8MiB    0.3%  33.8MiB
#   MTK model definition         1   8.67ms    0.0%  8.67ms    106KiB    0.0%   106KiB
# define callback                1    254ms    0.2%   254ms   8.89MiB    0.1%  8.89MiB
# ────────────────────────────────────────────────────────────────────────────────────
```

<details>
<summary>TTF script</summary>

```julia
using TimerOutputs
to = TimerOutput()
@timeit to "Import" begin
    using WorkshopCompanion
    using OpPoDyn
    using OpPoDyn.Library
    using ModelingToolkit
    using NetworkDynamics
    using Graphs
    using OrdinaryDiffEqRosenbrock
    using OrdinaryDiffEqNonlinearSolve
end

@timeit to "MTK Model creation" begin
    @timeit to "MTK model definition" begin
        @mtkmodel GenBus begin
            @components begin
                machine = SauerPaiMachine(;
                    vf_input=true,
                    τ_m_input=true,
                    S_b=100,
                    V_b=1,
                    ω_b=2π*60,
                    R_s=0.000125,
                    T″_d0=0.01,
                    T″_q0=0.01,
                    X_d, X′_d, X″_d, X_q, X′_q, X″_q, X_ls, T′_d0, T′_q0, H # free per machine parameter
                )
                avr = AVRTypeI(vr_min=-5, vr_max=5,
                    Ka=20, Ta=0.2,
                    Kf=0.063, Tf=0.35,
                    Ke=1, Te=0.314,
                    E1=3.3, Se1=0.6602, E2=4.5, Se2=4.2662,
                    tmeas_lag=false)
                gov = TGOV1(R=0.05, T1=0.05, T2=2.1, T3=7.0, DT=0, V_max=5, V_min=-5)
                busbar = BusBar()
            end
            @equations begin
                connect(machine.terminal, busbar.terminal)
                connect(machine.v_mag_out, avr.vh)
                connect(avr.vf, machine.vf_in)
                connect(gov.τ_m, machine.τ_m_in)
                connect(machine.ωout, gov.ω_meas)
            end
        end

        @mtkmodel LoadBus begin
            @components begin
                busbar = BusBar()
                load = ConstantYLoad(Pset, Qset)
            end
            @equations begin
                connect(load.terminal, busbar.terminal)
            end
        end
    end

    @timeit to "MTK model instntation" begin
        gen1p = (;machine__X_ls=0.01460, machine__X_d=0.1460, machine__X′_d=0.0608, machine__X″_d=0.06, machine__X_q=0.1000, machine__X′_q=0.0969, machine__X″_q=0.06, machine__T′_d0=8.96, machine__T′_q0=0.310, machine__H=23.64)
        gen2p = (;machine__X_ls=0.08958, machine__X_d=0.8958, machine__X′_d=0.1198, machine__X″_d=0.11, machine__X_q=0.8645, machine__X′_q=0.1969, machine__X″_q=0.11, machine__T′_d0=6.00, machine__T′_q0=0.535, machine__H= 6.40)
        gen3p = (;machine__X_ls=0.13125, machine__X_d=1.3125, machine__X′_d=0.1813, machine__X″_d=0.18, machine__X_q=1.2578, machine__X′_q=0.2500, machine__X″_q=0.18, machine__T′_d0=5.89, machine__T′_q0=0.600, machine__H= 3.01)

        @named mtkbus1 = GenBus(; gen1p...)
        @named mtkbus2 = GenBus(; gen2p...)
        @named mtkbus3 = GenBus(; gen3p...)
        @named mtkbus4 = MTKBus()
        @named mtkbus5 = LoadBus(;load__Pset=-1.25, load__Qset=-0.5)
        @named mtkbus6 = LoadBus(;load__Pset=-0.90, load__Qset=-0.3)
        @named mtkbus7 = MTKBus()
        @named mtkbus8 = LoadBus(;load__Pset=-1.0, load__Qset=-0.35)
        @named mtkbus9 = MTKBus()
    end
end

@timeit to "Bus building (simplify)" begin
    @named bus1 = Bus(mtkbus1; vidx=1, pf=pfSlack(V=1.04))
    @named bus2 = Bus(mtkbus2; vidx=2, pf=pfPV(V=1.025, P=1.63))
    @named bus3 = Bus(mtkbus3; vidx=3, pf=pfPV(V=1.025, P=0.85))
    @named bus4 = Bus(mtkbus4; vidx=4)
    @named bus5 = Bus(mtkbus5; vidx=5, pf=pfPQ(P=-1.25, Q=-0.5))
    @named bus6 = Bus(mtkbus6; vidx=6, pf=pfPQ(P=-0.9, Q=-0.3))
    @named bus7 = Bus(mtkbus7; vidx=7)
    @named bus8 = Bus(mtkbus8; vidx=8, pf=pfPQ(P=-1.0, Q=-0.35))
    @named bus9 = Bus(mtkbus9; vidx=9)
end

@timeit to "Line instantation" begin
    function piline(; R, X, B)
        @named pibranch = PiLine(;R, X, B_src=B/2, B_dst=B/2, G_src=0, G_dst=0)
        MTKLine(pibranch)
    end
    function transformer(; R, X)
        @named transformer = PiLine(;R, X, B_src=0, B_dst=0, G_src=0, G_dst=0)
        MTKLine(transformer)
    end

    @named l45 = Line(piline(; R=0.0100, X=0.0850, B=0.1760), src=4, dst=5)
    @named l46 = Line(piline(; R=0.0170, X=0.0920, B=0.1580), src=4, dst=6)
    @named l57 = Line(piline(; R=0.0320, X=0.1610, B=0.3060), src=5, dst=7)
    @named l69 = Line(piline(; R=0.0390, X=0.1700, B=0.3580), src=6, dst=9)
    @named l78 = Line(piline(; R=0.0085, X=0.0720, B=0.1490), src=7, dst=8)
    @named l89 = Line(piline(; R=0.0119, X=0.1008, B=0.2090), src=8, dst=9)
    @named t14 = Line(transformer(; R=0, X=0.0576), src=1, dst=4)
    @named t27 = Line(transformer(; R=0, X=0.0625), src=2, dst=7)
    @named t39 = Line(transformer(; R=0, X=0.0586), src=3, dst=9)
end

@timeit to "Network building" begin
    vertexfs = [bus1, bus2, bus3, bus4, bus5, bus6, bus7, bus8, bus9];
    edgefs = [l45, l46, l57, l69, l78, l89, t14, t27, t39];
    nw = Network(vertexfs, edgefs)
end

@timeit to "Solve powerflow" solve_powerflow!(nw)
@timeit to "Initialize" OpPoDyn.initialize!(nw)

@timeit to "define callback" begin
    deactivate_line = ComponentAffect([], [:pibranch₊active]) do u, p, ctx
        @info "Deactivate line $(ctx.src)=>$(ctx.dst) at t=$(ctx.t)"
        p[:pibranch₊active] = 0
    end
    cb = PresetTimeComponentCallback([1.0], deactivate_line)
    set_callback!(l46, cb)
end

@timeit to "Solve system" begin
    u0 = NWState(nw)
    prob = ODEProblem(nw, uflat(u0), (0,15), pflat(u0); callback=get_callbacks(nw))
    sol = solve(prob, Rodas5P())
end
show(to)
```

</details>